+++ CHANGELOG.md
@@ -2,5 +2,12 @@
 
 ## Unreleased
 
+### Breaking changes
+
 - rename `build/1` to `new/1`
 - rename `build!/1` to `new!/1`
+
+### New features
+
+- add `HTTPSpec.Request.build_method/1`
+- add `HTTPSpec.Request.build_url/1`